requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.9
class-registry==2.1.2
coreapi==2.3.3
coreschema==0.0.4
coverage==6.2
dj-database-url==0.5.0
Django==4.0
django-filter==21.1
django-on-heroku==1.1.2
djangorestframework==3.13.1
drf-yasg==1.20.0
filter==0.0.0.20200724
filters==1.3.2
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
packaging==21.3
psycopg2-binary==2.9.2
PyJWT==2.3.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
regex==2021.11.10
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.7
whitenoise==5.3.0
